Day4

+++ DAY 4/Playlist.cpp
@@ -10,22 +10,26 @@ int main() {
 
     int n;
     cin >> n;
-    vector<int> songs(n);
+    vector<long long> songs(n);
     
+    // Read song IDs
     for (int i = 0; i < n; ++i) {
         cin >> songs[i];
     }
 
-    unordered_set<int> unique_songs;
+    unordered_set<long long> unique_songs;
     int start = 0;
     int max_length = 0;
 
     for (int end = 0; end < n; ++end) {
+        // Remove elements from the start of the window until the current song is unique
         while (unique_songs.find(songs[end]) != unique_songs.end()) {
             unique_songs.erase(songs[start]);
             ++start;
         }
+        // Insert the current song into the set
         unique_songs.insert(songs[end]);
+        // Update the maximum length
         max_length = max(max_length, end - start + 1);
     }
